requirements.txt
boto3==1.40.74
botocore==1.40.74
certifi==2025.11.12
charset-normalizer==3.4.4
fsspec==2025.10.0
idna==3.11
jmespath==1.0.1
kafka-python==2.0.2
py4j==0.10.9.7
pyspark==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3fs==0.4.2
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
